fix(corfu): remove conflict with `+org/return`
Previously, when pressing Enter in org-mode, a newline would always be
inserted, regardless of corfu state. This led to annoying behavior: you
could select a candidate and preview it in the buffer via corfu, but RET
would erase its overlay and break the line instead of inserting the
candidate. This was due to a conflict with evil-org's binding.